Configure Hibernate framework

+++ scripts/create/tables/government-offices-table.sql
@@ -5,7 +5,7 @@ create table GOVERNMENT_OFFICES (
 	STATE                     varchar(50)    not null,
 	CITY                      varchar(50)    not null,
 	STREET                    varchar(50)    not null,
-	BUILDING_NUMBER           numeric(3, 0)  not null,
+	BUILDING_NUMBER           smallint       not null,
 	TELEPHONE_NUMBER          varchar(20)    not null,
 
 	primary key (GOVERNMENT_OFFICE_ID)
